requirements.txt
#common
scikit-image
# local install
#paddlepaddle-gpu==2.2.0
# kaggle competitions download -c imagenet-object-localization-challenge
#kaggle
# CPU
paddlepaddle
opencv-python
tqdm
